Restructure testing databases like the models
