+++ packages/data/testing/db/mock/api/index.ts
@@ -1,0 +1,3 @@
+export * from './ability.db';
+export * from './race.db';
+export * from './skill.db';
