+++ packages/data/testing/db/mock/authentication/index.ts
@@ -1,0 +1,2 @@
+export * from './user-role.db';
+export * from './user.db';
