+++ packages/data/testing/index.ts
@@ -1,5 +1,1 @@
-export * from './db/mock/ability.db';
-export * from './db/mock/race.db';
-export * from './db/mock/skill.db';
-export * from './db/mock/user-role.db';
-export * from './db/mock/user.db';
+export * from './db/mock';
